+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 
 v0.0.2-dev
 ----------
++ Add link to stringDB
 + Add link to gnomAD
 + Add API endpoint /links/ to fetch URL's for external references
 + Fetch knownGene track from UCSC