+++ CHANGELOG.rst
@@ -6,8 +6,12 @@ Changelog
 .. This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
-v0.0.2-dev
+v0.0.3-dev
 ----------
+
+v0.0.2
+------
++ Add FastAPI endpoint for transcripts and links
 + Add link to stringDB
 + Add link to gnomAD
 + Add API endpoint /links/ to fetch URL's for external references
@@ -15,7 +19,7 @@ v0.0.2-dev
 + Fetch transcript information from Ensembl
 
 v0.0.1
-----------
+------
 + Add class for Transcripts
 + Validate that the input is valid Bed format
 + Add class for Bed format